+++ projects/demo/src/app/app.component.ts
@@ -40,6 +40,17 @@ import {
     DistributePageComponent,
     CardPageComponent
 } from './demo';
+import {
+    TableSortableHeaderPageComponent
+} from './demo/pages/table-page/table-sortable-header-page/table-sortable-header-page.component';
+import { TableRowPageComponent } from './demo/pages/table-page/table-row/table-row-page.component';
+import { TableCellPageComponent } from './demo/pages/table-page/table-cell-page/table-cell-page.component';
+import {
+    TableSortableHeaderCellPageComponent
+} from './demo/pages/table-page/table-sortable-header-cell-page/table-sortable-header-cell-page.component';
+import {
+    TableLeadCellMultiLinePageComponent
+} from './demo/pages/table-page/table-lead-cell-multi-line-page/table-lead-cell-multi-line-page.component';
 
 export const menu: IDemoMenuSection[] = [
     {
@@ -75,12 +86,22 @@ export const menu: IDemoMenuSection[] = [
             { view: 'popup', title: 'Popup', type: PopupPageComponent },
             { view: 'popover', title: 'Popover', type: PopoverPageComponent },
             { view: 'sidenav', title: 'Sidenav', type: SidenavPageComponent },
-            { view: 'tables', title: 'Tables', type: TablePageComponent },
             { view: 'tabs', title: 'Tabs', type: TabsPageComponent },
             { view: 'toast', title: 'Toast', type: ToastPageComponent },
             { view: 'tooltip', title: 'Tooltip', type: TooltipPageComponent },
             { view: 'tooltip-ellipsis', title: 'Ellipsis tooltip', type: EllipsisTooltipPageComponent },
         ],
+    },
+    {
+        title: 'Tables',
+        pages: [
+            { view: 'table', title: 'Table', type: TablePageComponent },
+            { view: 'table-row', title: 'Table row', type: TableRowPageComponent },
+            { view: 'table-cell', title: 'Table cell', type: TableCellPageComponent },
+            { view: 'table-lead-cell-multi-line', title: 'Table lead cell multi line', type: TableLeadCellMultiLinePageComponent },
+            { view: 'table-sortable-header', title: 'Table sortable header', type: TableSortableHeaderPageComponent },
+            { view: 'table-sortable-header-cell', title: 'Table sortable header cell', type: TableSortableHeaderCellPageComponent },
+        ]
     },
     {
         title: 'Form elements',